+++ packages/client/base/src/types/embed/crypto.ts
@@ -1,7 +1,7 @@
 import type { Transaction as _EthersTransaction } from "@ethersproject/transactions";
 import type { Transaction as _SolanaTransaction } from "@solana/web3.js";
 
-import type { EVMBlockchainIncludingTestnet, SignerSupportedBlockchains } from "@crossmint/common-sdk-base";
+import type { EVMBlockchainIncludingTestnet } from "@crossmint/common-sdk-base";
 
 import type { CommonEmbeddedCheckoutProps } from ".";
 import { CryptoPaymentMethod } from "..";
@@ -54,20 +54,4 @@ export type ETHEmbeddedCheckoutSigner = CommonEmbeddedCheckoutSignerProps & {
 
 export type SOLEmbeddedCheckoutSigner = CommonEmbeddedCheckoutSignerProps & {
     signAndSendTransaction: (transaction: SolanaTransaction) => Promise<string>;
-};
-
-export type TransactionResponse =
-    | {
-          success: true;
-          txId: string;
-      }
-    | {
-          success: false;
-          errorMessage: string;
-      };
-
-export type EmbeddedCheckoutSigner = {
-    address: string;
-    chain: SignerSupportedBlockchains;
-    handleSignAndSendTransaction(serializedTransaction: string): Promise<TransactionResponse>;
 };
