+++ packages/client/base/src/types/embed/v3/CrossmintEmbeddedCheckoutV3Props.ts
@@ -1,7 +1,7 @@
-import type { EmbeddedCheckoutSigner, Currency as FiatCurrency, Locale } from "@/types";
+import type { Currency as FiatCurrency, Locale } from "@/types";
 import type { CryptoCurrency } from "@/types/CryptoCurrency";
 
-import type { BlockchainIncludingTestnet } from "@crossmint/common-sdk-base";
+import type { BlockchainIncludingTestnet, PayerSupportedBlockchains } from "@crossmint/common-sdk-base";
 
 interface CrossmintEmbeddedCheckoutV3CommonProps {
     appearance?: EmbeddedCheckoutV3Appearance;
@@ -190,7 +190,22 @@ export type EmbeddedCheckoutV3CryptoPayment = {
     enabled: boolean;
     defaultChain?: BlockchainIncludingTestnet;
     defaultCurrency?: CryptoCurrency;
-    signer?: EmbeddedCheckoutSigner;
+    payer?: EmbeddedCheckoutPayer;
     // allowedCurrencies?: Partial<Record<BlockchainIncludingTestnet, false | CryptoCurrency[]>>; // TODO: Add this back when supported on crossmint-main
-    // payer?: any; // TODO: Add this back when supported on crossmint-main
+};
+
+export type TransactionResponse =
+    | {
+          success: true;
+          txId: string;
+      }
+    | {
+          success: false;
+          errorMessage: string;
+      };
+
+export type EmbeddedCheckoutPayer = {
+    address: string;
+    chain: PayerSupportedBlockchains;
+    handleSignAndSendTransaction(serializedTransaction: string): Promise<TransactionResponse>;
 };
